requirements.txt
babel==2.6.0
certifi==2019.3.9
chaostoolkit-lib==1.1.2
chardet==3.0.4
debtcollector==1.21.0
idna==2.8
iso8601==0.1.12
keystoneauth1==3.13.1
logzero==1.5.0
msgpack==0.6.1
netaddr==0.7.19
netifaces==0.10.9
os-service-types==1.6.0
oslo.i18n==3.23.1
oslo.serialization==2.28.2
oslo.utils==3.40.3
pbr==5.1.3
prettytable==0.7.2
pycodestyle==2.5.0
pyparsing==2.3.1
python-novaclient==13.0.0
pytz==2018.9
pyyaml==5.1
requests==2.21.0
simplejson==3.16.0
six==1.12.0
stevedore==1.30.1
urllib3==1.24.1
wrapt==1.11.1
